enhance background activity service test coverage
