+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/FakeSession.kt
@@ -3,7 +3,6 @@ package io.embrace.android.embracesdk.fakes
 import io.embrace.android.embracesdk.payload.Session
 import io.embrace.android.embracesdk.payload.Session.Companion.APPLICATION_STATE_FOREGROUND
 import io.embrace.android.embracesdk.payload.SessionMessage
-import io.embrace.android.embracesdk.session.MESSAGE_TYPE_START
 
 internal fun fakeSession(): Session = Session(
     sessionId = "fakeSessionId",
@@ -13,7 +12,7 @@ internal fun fakeSession(): Session = Session(
     isColdStart = true,
     startType = Session.LifeEventType.STATE,
     properties = mapOf(),
-    messageType = MESSAGE_TYPE_START,
+    messageType = Session.MESSAGE_TYPE_END,
     user = null
 )
