+++ embrace-android-sdk/src/test/resources/breadcrumb_custom.json
@@ -3,7 +3,7 @@
     "id": "fakeSessionId",
     "st": 160000000000,
     "sn": 1,
-    "ty": "st",
+    "ty": "en",
     "as": "foreground",
     "cs": true,
     "sm": "s",
