+++ embrace-android-sdk/src/test/resources/session_stacktrace_symbols.json
@@ -2,7 +2,7 @@
   "id": "fakeSessionId",
   "st": 160000000000,
   "sn": 1,
-  "ty": "st",
+  "ty": "en",
   "as": "foreground",
   "cs": true,
   "sm": "s",
